# add faster, less allocatey unicode escaper