Breaking change - regex version supported invalid format `\UNNNN`, this version requires `\U` prefix for 8 digit version `\UNNNNNNNN` and `\u` prefix for 4 digits `\uNNNN`


Benchmark for single replacement:

  Method |  Job | Runtime |        Mean |      Error |     StdDev | Ratio | RatioSD | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-------- |----- |-------- |------------:|-----------:|-----------:|------:|--------:|-----:|------------:|------------:|------------:|--------------------:|
   Regex |  Clr |     Clr | 1,048.77 ns | 20.7532 ns | 40.9648 ns |  1.00 |    0.00 |    4 |      0.3090 |           - |           - |               976 B |
 Loop_16 |  Clr |     Clr |    36.85 ns |  0.8045 ns |  0.9577 ns |  0.03 |    0.00 |    1 |      0.0228 |           - |           - |                72 B |
 Loop_32 |  Clr |     Clr |    37.67 ns |  0.8440 ns |  1.5433 ns |  0.04 |    0.00 |    1 |      0.0228 |           - |           - |                72 B |
   Regex | Core |    Core |   856.17 ns | 12.0266 ns | 11.2497 ns |  0.80 |    0.04 |    3 |      0.1974 |           - |           - |               624 B |
 Loop_16 | Core |    Core |    39.20 ns |  0.8530 ns |  1.5382 ns |  0.04 |    0.00 |    2 |      0.0228 |           - |           - |                72 B |
 Loop_32 | Core |    Core |    37.41 ns |  0.8348 ns |  0.7400 ns |  0.04 |    0.00 |    1 |      0.0228 |           - |           - |                72 B |

Benchmark for 50 replacements:

  Method |  Job | Runtime |      Mean |     Error |    StdDev | Ratio | RatioSD | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
-------- |----- |-------- |----------:|----------:|----------:|------:|--------:|-----:|------------:|------------:|------------:|--------------------:|
   Regex |  Clr |     Clr | 33.824 us | 0.3599 us | 0.3191 us |  1.00 |    0.00 |    2 |      8.0566 |           - |           - |             25521 B |
   Regex | Core |    Core | 32.855 us | 0.7109 us | 0.6302 us |  0.97 |    0.02 |    1 |      7.8125 |           - |           - |             24728 B |
 Loop_16 |  Clr |     Clr |  1.080 us | 0.0210 us | 0.0327 us |  1.00 |    0.00 |    1 |      0.3014 |           - |           - |               952 B |
 Loop_16 | Core |    Core |  1.118 us | 0.0239 us | 0.0392 us |  1.04 |    0.05 |    2 |      0.3014 |           - |           - |               952 B |
 Loop_32 |  Clr |     Clr |  1.988 us | 0.0409 us | 0.0471 us |  1.00 |    0.00 |    1 |      0.4616 |           - |           - |              1456 B |
 Loop_32 | Core |    Core |  1.987 us | 0.0389 us | 0.0532 us |  1.00 |    0.04 |    1 |      0.4616 |           - |           - |              1456 B |