@msigut 2018-11-26
Great work Lukas :) As we agreed together, logic is now carry by Newtonsoft.JSON library.